+++ libubuntudownloadmanager/download.h
@@ -76,10 +76,10 @@ public slots:
     // slots that are exposed via dbus, they just change the state, the downloader
     // takes care of the actual download operations
     QVariantMap metadata();
-    uint progress();
-    uint totalSize();
-    virtual void setThrottle(uint speed);
-    virtual uint throttle();
+    qlonglong progress();
+    qlonglong totalSize();
+    virtual void setThrottle(qlonglong speed);
+    virtual qlonglong throttle();
     void allowGSMDownload(bool allowed);
     bool isGSMDownloadAllowed();
     void cancel();
@@ -93,7 +93,7 @@ signals:
     void error(const QString& error);
     void finished(const QString& path);
     void paused(bool success);
-    void progress(uint received, uint total);
+    void progress(qlonglong received, qlonglong total);
     void resumed(bool success);
     void started(bool success);
